Heavily refactor exception handling and remove dependence on CMSIS
The exception handling code has been heavily refactored to provide more
flexibility to application developers around configuring interrupts, but
also to provide more information when an internal exception occurs. In
those cases, the register state of the crashing thread will be preserved
on the main stack and dumped to the log for inspection.

Dependence on CMSIS was also removed by writing an NVIC driver and
cache invalidation/enablement code.

Lastly, a process stack was allocated and initialized for the initial
thread and setup during the Reset handler. This is in preparation to
using the initial thread as an idle thread when a scheduler gets
written.